FechaCompleta, imágenes que has insertado y preguntar antes de eliminar

+++ Angular/equipos/src/app/listado-equipos/listado-equipos.component.ts
@@ -1,38 +1,55 @@
 import { Component, OnInit } from '@angular/core';
 import { BackendService } from '../services/backend.service';
 import Swal from 'sweetalert2';
+import { registerLocaleData } from '@angular/common';
+import localeEs from '@angular/common/locales/es';
 
 @Component({
   selector: 'app-listado-equipos',
   templateUrl: './listado-equipos.component.html',
-  styleUrls: ['./listado-equipos.component.css']
+  styleUrls: ['./listado-equipos.component.css'],
 })
 export class ListadoEquiposComponent implements OnInit {
-  equipos:any;
+  equipos: any;
 
-  constructor(private backend:BackendService) { }
+  constructor(private backend: BackendService) {}
 
   ngOnInit(): void {
     this.backend.getEquipos().subscribe((equipos) => {
       this.getEquipos();
-    })
+    });
+    registerLocaleData(localeEs);
+    this.getEquipos();
   }
   //ordena por nombre
-  getEquipos(){
+  getEquipos() {
     this.backend.getEquipos().subscribe((equipos) => {
       this.equipos = equipos;
       this.equipos.sort((a: any, b: any) => a.nombre.localeCompare(b.nombre));
       console.log(this.equipos);
-    })
+    });
   }
-
-  deleteEquipo(equipo:any){
-    this.backend.deleteEquipos(equipo.id).subscribe(() =>{
+  esImagen(nombre: string) {
+    return nombre.startsWith('data:image/png;base64');
+  }
+  deleteEquipo(equipo: any) {
+    this.backend.deleteEquipos(equipo.id).subscribe(() => {
       Swal.fire({
-        icon: 'success',
-        title: 'Registro borrado',
-        text: `El equipo ${equipo.nombre} se ha borrado correctamente`
-      }).then(() => {this.getEquipos();})
+        icon: 'question',
+        title: '¿Estás seguro?',
+        showCancelButton: true,
+        cancelButtonText: 'Cancelar',
+      }).then((result) => {
+        if (result.isConfirmed) {
+          this.backend.deleteEquipos(equipo.id).subscribe(() => {
+            this.getEquipos();
+            Swal.fire({
+              icon: 'success',
+              title: `El jugador '${equipo.nombre}' se ha borrado correctamente`,
+            });
+          });
+        }
+      });
     });
   }
 }
