+++ Angular/equipos/src/app/listado-jugadores/listado-jugadores.component.ts
@@ -5,34 +5,41 @@ import { BackendService } from '../services/backend.service';
 @Component({
   selector: 'app-listado-jugadores',
   templateUrl: './listado-jugadores.component.html',
-  styleUrls: ['./listado-jugadores.component.css']
+  styleUrls: ['./listado-jugadores.component.css'],
 })
 export class ListadoJugadoresComponent implements OnInit {
+  jugadores: any;
 
-  jugadores:any;
-
-  constructor(private backend:BackendService) { }
+  constructor(private backend: BackendService) {}
 
   ngOnInit(): void {
-      this.getjugadores();
+    this.getjugadores();
   }
   //ordena por nombre
-  getjugadores(){
+  getjugadores() {
     this.backend.getJugadores().subscribe((jugadores) => {
       this.jugadores = jugadores;
       this.jugadores.sort((a: any, b: any) => a.numero - b.numero);
       console.log(this.jugadores);
-    })
-  }
-
-  deleteJugadores(jugador:any){
-    this.backend.deleteJugadores(jugador.id).subscribe(() =>{
-      Swal.fire({
-        icon: 'success',
-        title: 'Registro borrado',
-        text: `El jugador ${jugador.nombre} se ha borrado correctamente`
-      }).then(() => {this.getjugadores();})
     });
   }
 
+  deleteJugadores(jugador: any) {
+    Swal.fire({
+      icon: 'question',
+      title: '¿Estás seguro?',
+      showCancelButton: true,
+      cancelButtonText: 'Cancelar',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.backend.deleteJugadores(jugador.id).subscribe(() => {
+          this.getjugadores();
+          Swal.fire({
+            icon: 'success',
+            title: `El jugador '${jugador.nombre}' se ha borrado correctamente`,
+          });
+        });
+      }
+    });
+  }
 }
